Fix v-sync mode handling
After playtesting the game on GMS2, I realized that restarting the game after toggling the option is useless. The displaying bugs only happen in GMS1.4. While GMS2 handles it flawlessly, a fix had to be made as the game also gets exported to said GameMaker version.

An additional fix was also applied to the control handling for the Android version.